requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dash==2.16.1
dash-bootstrap-components==1.6.0
dash-bootstrap-templates==1.1.2
dash-core-components==2.0.0
dash-credit-cards==0.0.1
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-mantine-components==0.12.1
dash-table==5.0.0
dill==0.3.8
diskcache==5.6.3
dnspython==2.6.1
email_validator==2.1.1
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Redmail==0.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
future==0.15.2
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
multiprocess==0.70.16
nest-asyncio==1.6.0
numpy==1.26.4
openpay @ git+https://github.com/Simon-U/openpay-fork.git@3b4dee0c724331fb4cdf9b9b12e95e12e9608dd6
packaging==24.0
pandas==2.2.2
plotly==5.21.0
polars==0.20.21
psutil==5.9.8
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redmail==0.6.0
requests==2.31.0
retrying==1.3.4
six==1.16.0
SQLAlchemy==2.0.29
tenacity==8.2.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
WTForms==3.1.2
zipp==3.18.1
